added secure file deletion, overwriting the file with random data before actual deletion
Update issue 56
Added a method for secure file deletion. We'll possibly have to test this a bit and make sure it actually *is* secure on the filesystems used (vfat for SD card, yaffs2 for phone flash memory by default). This sort of overwrite might be absolutely useless on some filesystems.
I also wanted to add an option at first, but this seems reasonably fast and in general it should be what the user wants, so for now it will ALWAYS securely delete. If there is a demand for plain deletion, then options can be added.

Update issue 39
A new string:
<string name="progress_deletingSecurely">deleting \'%s\' securely...</string>