Normalize public key uri in KeyListFragment
Find the explicit row id of the public key id given in the uri in
KeyListFragment and work with that. This way, passing in uris by
master key id, mail address, or any other criteria, works.

Might be a good idea to add an actual check if the row id is non-zero
here, but not sure how to do a "bad intent" thing