A more real fix for #1077
Ticket #1077 is a horrible bug. It's multi-faceted. It's a *fractal* bug.

ConsolidateDialogActivity is being created more than once. Don't ask me why or
for what reason, but it is. Someone, please tell me what's really going on some
day when I'm drinking hard liquor. I'd like to know.

Changes in KeychainIntentService are:

 * Clean up the logic. We can spare a line on the stack, and it's a cleaner
 approach to problem there.

Changes in KeyListFragment are:

 * Consolidate helper function calls into ConsolidateDialogActivity such that
 it's a reproducable problem, and avoids code duplication.

AndroidManifest is where the "meat" of this patch lives. When you declare an
activity as `android:launchMode="singleInstance"`, it becomes a slightly
magical activity. If there is no instance of the activity already, `onCreate`
is called and things are done to make sure it works right. However, if there is
already a running instance of the class specified, Android calls it's
`onNewIntent` function. Because `ConsolidateDialogActivity` has no `onNewIntent`
 the super's is called - which is empty.

So everyone's happy! There's a deeper problem somewhere, but this fixes the
immediate problem: the intent that spawns `ConsolidateDialogActivity` a
second time is being squashed.