display unknown keys red rather than orange
For an unknown key, there is no indication of the state the key is in.
To indicate both immediate action required, and to make this status
equal to its worst case (rather than *better* than its worst case), the
status is displayed in red.

At some point, we will probably want to download unknown keys
automatically, at which point an unknown key will actually be an error
state. This is an intermediate solution until then.