automatically open decrypted+verified files if default viewer is set
In standard Android UX, the user can set a default viewer app using the
"Always" button in the app chooser.  After that, the expectation is that
clicking that file will view the app.  If the user decrypts a file, they
most likely want to do view it, not just see whether it decrypts. So this
makes it automatically view the decrypted file if the signature has been
verified.  That last step keeps it safe, so that OpenKeychain must trust
the creator of the file before opening potential malware. If the user did
not want to view the file, they can click Back and they'll be on the
normal OpenKeychain screen, where they can choose Share, Save, etc.

Once OpenKeychain does not automatically decrypt to disk, then this should
probably be modified to only automatically view using apps that support
streaming rather than direct file access.