Fix digest handling when signing using a hardware token
The previous code used java.security.MessageDigest.digest()
more than once, which is incorrect because the digest is reset
after each call.

This made OpenKeychain generate invalid signatures when
using a hardware token: the two bytes called "fingerprint",
which are supposed to be the first two bytes of the unsigned
digest, were incorrectly filled inside the signature.

Some tools do not check this fingerprint, only the signed digest,
so the signature was reported as valid. GPG is one of these tools.

But some other tools (like OpenPGP.js which uses rnp) did check it,
and reported an invalid signature, assorted with an error message
saying "wrong lbits". Thunderbird is one of the users of OpenPGP.js

This fixes bug #2688